streetfighter/requirements.txt
gymnasium==0.29.1
numpy==1.26.3
opencv_python==4.8.1.78
optuna==3.5.0
stable_baselines3==2.2.1
stable_retro==0.9.2
